restructure core for crates folder
- improve naming consistency
- added placeholder files for near future code
- added docs
