+++ .cspell/project_words.txt
@@ -1,42 +1,46 @@
-spacedrive
-spacedriveapp
-vdfs
-haoyuan
-brendonovich
-codegen
-elon
-deel
-haden
 akar
-benja
-haris
-mehrzad
-OSSC
-josephjacks
-rauch
-ravikant
-neha
-narkhede
 allred
-lütke
-tobiaslutke
-justinhoffman
-rywalker
-zacharysmith
-sanjay
-poonen
-mytton
-davidmytton
-richelsen
-lesterlee
 alluxio
 augusto
+benja
+brendonovich
+chacha
+codegen
+davidmytton
+deel
+elon
+encryptor
+haden
+haoyuan
+haris
+josephjacks
+justinhoffman
+Keyslot
+keyslots
+lesterlee
+lütke
 marietti
-vijay
-sharma
+mehrzad
+mytton
+narkhede
 naveen
+neha
 noco
+OSSC
+poonen
+rauch
+ravikant
+richelsen
 rspc
 rspcws
-stringly
+rywalker
+sanjay
+sharma
+spacedrive
+spacedriveapp
 specta
+stringly
+tobiaslutke
+vdfs
+vijay
+zacharysmith
