--- .vscode/.todo
@@ -1,50 +1,0 @@
-####
-# Completed
-####
-✔ Configured a Tauri window 
-✔ Primative UI components (Button, Input, Shortcut etc) 
-✔ Connect & create sqlite database in application data folder
-✔ Auto migrate database on app launch 
-✔ Define data schema in Rust 
-✔ Scan given directory recursively and write to database 
-✔ App frame and router set up 
-✔ Render basic file list with database data 
-✔ Create a global store 
-✔ Virtualized list for explorer @done(21-10-24 04:03)
-    Also cache folder contents and serve lazy, updating in background
-✔ Native file type icons @done(21-10-24 04:04)
-✔ Cache native icons in app data folder @done(21-10-24 11:30)
-✔ File inspector @done(21-10-24 23:30)
-✔ Upgrade to mono-repo @done(21-12-27 18:30)
-✔ Create project readme @done(21-12-28 18:30)
-✔ Refactor indexer to scan efficiently @done(21-12-29 18:30)
-✔ File previews
-✔ Generate CAS ids
-✔ Settings screen
-✔ Job queue system
-✔ Landing page
-✔ Volume identification
-✔ Statistic calucations
-    Folder size, Volume size, Total capacity, Total unique etc..
-    
-####
-# Next up 
-####
-☐ Tag creation and assignment
-☐ Right click menu
-☐ Set up Tauri updater
-☐ Onboarding flow
-☐ Explorer grid view
-☐ Changelog
-
-####
-# Planned features
-####
-☐ Action system & command pallete
-☐ Explorer path viewer / editor
-☐ File encryptor
-☐ File viewer / player
-☐ Secret keystore
-☐ Search 
-    Efficient way to search sqlite: make file table WITHOUT ROWID
-☐ Open with
