+++ core/Cargo.toml
@@ -16,7 +16,7 @@ mobile = [
 ] # This feature allows features to be disabled when the Core is running on mobile.
 ffmpeg = [
   "dep:ffmpeg-next",
-  "dep:thumbnailer",
+  "dep:sd_ffmpeg",
 ] # This feature controls whether the Spacedrive Core contains functionality which requires FFmpeg.
 
 [dependencies]
@@ -56,7 +56,8 @@ async-trait = "^0.1.52"
 image = "0.24.1"
 webp = "0.2.2"
 ffmpeg-next = { version = "5.0.3", optional = true, features = [] }
-thumbnailer = { path = "./thumbnailer", optional = true }
+sd_ffmpeg = { path = "../crates/ffmpeg", optional = true }
+crypto = { path = "../crates/crypto" }
 fs_extra = "1.2.0"
 tracing = "0.1.35"
 tracing-subscriber = { version = "0.3.14", features = ["env-filter"] }
