--- core/package.json
@@ -1,17 +1,0 @@
-{
-	"name": "@sd/core",
-	"version": "0.0.0",
-	"main": "index.js",
-	"license": "GPL-3.0-only",
-	"scripts": {
-		"codegen": "cargo test",
-		"build": "cargo build",
-		"test": "cargo test",
-		"test:log": "cargo test -- --nocapture",
-		"prisma": "cargo prisma"
-	},
-	"devDependencies": {
-		"@types/node": "^18.6.1",
-		"typescript": "^4.7.4"
-	}
-}
