+++ core/src/api/jobs.rs
@@ -1,8 +1,10 @@
 use crate::{
-	encode::{ThumbnailJob, ThumbnailJobInit},
-	file::cas::{FileIdentifierJob, FileIdentifierJobInit},
 	job::{Job, JobManager},
 	location::{fetch_location, LocationError},
+	object::{
+		identifier_job::{FileIdentifierJob, FileIdentifierJobInit},
+		preview::{ThumbnailJob, ThumbnailJobInit},
+	},
 	prisma::location,
 };
