--- core/src/file/cas/mod.rs
@@ -1,5 +1,0 @@
-mod checksum;
-mod identifier;
-
-pub use checksum::*;
-pub use identifier::*;
