+++ core/src/lib.rs
@@ -13,16 +13,15 @@ use tracing::{error, info};
 use tracing_subscriber::{filter::LevelFilter, fmt, prelude::*, EnvFilter};
 
 pub mod api;
-pub(crate) mod encode;
-pub(crate) mod file;
 pub(crate) mod job;
 pub(crate) mod library;
 pub(crate) mod location;
 pub(crate) mod node;
 pub(crate) mod object;
-pub(crate) mod prisma;
 pub(crate) mod util;
 pub(crate) mod volume;
+
+pub(crate) mod prisma;
 
 #[derive(Clone)]
 pub struct NodeContext {
