+++ core/src/location/indexer/rules.rs
