+++ core/src/object/mod.rs
@@ -1,3 +1,7 @@
+pub mod cas;
+pub mod identifier_job;
+pub mod preview;
+
 // Objects are primarily created by the identifier from Paths
 // Some Objects are purely virtual, unless they have one or more associated Paths, which refer to a file found in a Location
 // Objects are what can be added to Spaces
