+++ core/src/object/preview/metadata.rs
