+++ crates/ffmpeg/src/error.rs
