+++ crates/ffmpeg/src/film_strip.rs
