+++ crates/ffmpeg/src/thumbnailer.rs
