+++ crates/p2p/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+	println!("Hello, world!");
+}
