+++ crates/prisma/.gitignore
@@ -1,0 +1,1 @@
+/src/prisma.rs
