+++ docs/product/guides/adding-locations.md
@@ -4,4 +4,39 @@ index: 5
 
 # Adding Locations
 
-![image](/locations.webp)
+Locations are places Spacedrive looks for files. You can add Locations from any device to your Library, they will be scanned and monitored for filesystem changes.
+
+![Spacedrive Locations from the Settings view](/locations.webp)
+
+Locations can be managed from the settings. The "online" indicator shows if that Location is currently accessible from one of your devices.
+
+## Scanning
+When a Location is added it is immediately scanned. The scan happens via several phases.
+
+1. The [Indexer](): directory structure is indexed and saved to your library's database.
+2. The [Identifier](): metadata is extracted and unique identifiers generated.
+3. The [Thumbnailer](): compressed previews are generated for all media and text files.
+
+If any one of these jobs fail or the application is closed, the state is saved and it will resume when the app is relaunched. 
+
+::: slot info 
+If you wish to pause or cancel these jobs you can do so from the [Job Manager]()
+:::
+
+There is a button to trigger a full re-scan of this Location. Otherwise the location will keep watch for filesystem changes automatically.
+
+Deleting a location will remove the data from the database permanently.
+
+## Archiving
+::: slot warning 
+This feature is WIP
+:::
+Locations can be archived, meaning the directory structure is extracted from the Spacedrive database and preserved as a standalone database file. This file can be opened by Spacedrive in the [Database]() view.
+
+Archives are useful when files have been moved or deleted from a Location, but you want to keep a record of the directory structure of that Location without cluttering the Spacedrive interface. File metadata can optionally be included/excluded from the archive file.
+
+## Dynamic Space
+A slider in the Location settings allows you to allocate an amount of GB to use as redundancy for other Locations. The storage space will be filled and emptied dynamically, recent and important files will be copied as priority.
+
+## Encrypted Space
+Allocate a portion of this location to be encrypted
