+++ docs/product/guides/connecting-nodes.md
@@ -3,3 +3,12 @@ index: 6
 ---
 
 # Connecting Nodes
+
+::: slot note WIP 
+It is not possible to connect Nodes yet, please check our [roadmap](/roadmap).
+:::
+
+Nodes are instances of Spacedrive, commonly your devices. Currently Spacedrive supports Windows, Linux, Mac, iOS and Android. 
+
+These devices can be connected via P2P to directly and securely synchronize and move data. 
+
