+++ docs/product/guides/creating-spaces.md
@@ -3,3 +3,5 @@ index: 7
 ---
 
 # Creating Spaces
+
+Spaces are virtual folders, they can contain files from various locations.
