--- docs/product/guides/database-backup.md
@@ -1,5 +1,0 @@
----
-index: 6
----
-
-# Database Backup
