+++ docs/product/guides/find-my-file.md
@@ -1,0 +1,7 @@
+---
+index: 10
+---
+
+# Find My File
+
+Mark a file as lost, if Spacedrive ever discovers that file when indexing, you will be notified and Spacedrive will make a backup.
