+++ docs/product/guides/library-setup.md
@@ -1,0 +1,14 @@
+---
+index: 0
+---
+
+# Library Setup
+
+A Library is where all your Spacedrive data lives, but not the files themselves.
+
+For the convenience of finding, organizing and sharing files even when you don't have access; directory structures, metadata and preview media are extracted and stored within the Spacedrive Library.
+
+This data is heavily compressed and very portable, you can have multiple libraries and even lock and encrypt them.
+
+## Deleting a Library
+A Library can be deleted,
