+++ package.json
@@ -3,11 +3,11 @@
 	"version": "0.0.0",
 	"private": true,
 	"scripts": {
-		"prep": "pnpm db:gen && pnpm core codegen",
+		"prep": "pnpm db:gen && cargo test",
 		"build": "turbo run build",
 		"landing-web": "turbo run dev --parallel --filter=@sd/landing --filter=@sd/web",
-		"db:migrate": "pnpm core prisma migrate dev",
-		"db:gen": "pnpm core prisma generate",
+		"db:migrate": "cd crates && cargo prisma migrate dev",
+		"db:gen": "cd crates && cargo prisma generate",
 		"format": "prettier --config .prettierrc.cli.js --write \"**/*.{ts,tsx,html,scss,json,yml,md}\"",
 		"desktop": "pnpm --filter @sd/desktop --",
 		"web": "pnpm --filter @sd/web -- ",
